rename: 프로젝트 패키지 구조 변경 혼동되는 패키지가 있어서 수정합니다.
- xenum -> 각 도메인의 entity로 이동
- security 패키지를 삭제하고, token과 global 패키지로 분리
- global/exception -> 각 도메인의 exception으로 이동
