+++ src/main/java/com/prgrms2/java/bitta/global/config/JpaAuditConfig.java
@@ -1,4 +1,4 @@
-package com.prgrms2.java.bitta.config;
+package com.prgrms2.java.bitta.global.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
