+++ src/main/java/com/prgrms2/java/bitta/jobpost/entity/PayStatus.java
@@ -1,4 +1,4 @@
-package com.prgrms2.java.bitta.xenum;
+package com.prgrms2.java.bitta.jobpost.entity;
 
 public enum PayStatus {
     FREE, PAID // 모집 방식 (무급, 유급)
