+++ src/main/java/com/prgrms2/java/bitta/user/entity/CustomUserPrincipal.java
@@ -1,4 +1,4 @@
-package com.prgrms2.java.bitta.security.auth;
+package com.prgrms2.java.bitta.user.entity;
 
 import lombok.RequiredArgsConstructor;
 
@@ -7,7 +7,6 @@ import java.security.Principal;
 @RequiredArgsConstructor
 public class CustomUserPrincipal implements Principal {
     private final String email;
-
 
     @Override
     public String getName() {
