+++ src/main/java/com/prgrms2/java/bitta/user/entity/Role.java
@@ -1,4 +1,4 @@
-package com.prgrms2.java.bitta.xenum;
+package com.prgrms2.java.bitta.user.entity;
 
 public enum Role {
     Actor,
